renamed model dirs to lower case
