+++ src/builders/AvailabilityCalendarModelBuilder.ts
@@ -2,8 +2,8 @@ import { AvailabilityStatus, CapabilityId, PricingPer } from "@octocloud/types";
 import { PricingDataProvider } from "../dataProviders/PricingDataProvider";
 import { DateHelper } from "../helpers/DateHelper";
 import { AvailabilityCalendarData } from "../data/AvailabilityCalendarData";
-import { AvailabilityCalendarModel } from "../models/Availability/AvailabilityCalendarModel";
-import { AvailabilityCalendarPricingModel } from "../models/Availability/AvailabilityCalendarPricingModel";
+import { AvailabilityCalendarModel } from "../models/availability/AvailabilityCalendarModel";
+import { AvailabilityCalendarPricingModel } from "../models/availability/AvailabilityCalendarPricingModel";
 import { AvailabilityCalendarPricingModelFactory } from "../factories/AvailabilityCalendarPricingModelFactory";
 
 interface AvailabilityCalendarModelBuilderData {
