+++ src/builders/OptionModelBuilder.ts
@@ -1,14 +1,14 @@
 import { CapabilityId, DurationUnit, PricingPer } from "@octocloud/types";
 import { UnitModelBuilder } from "./UnitModelBuilder";
 import { OptionData } from "../data/OptionData";
-import { OptionModel } from "../models/Option/OptionModel";
-import { OptionContentModel } from "../models/Option/OptionContentModel";
-import { OptionPickupModel } from "../models/Option/OptionPickupModel";
-import { OptionPricingModel } from "../models/Option/OptionPricingModel";
+import { OptionModel } from "../models/option/OptionModel";
+import { OptionContentModel } from "../models/option/OptionContentModel";
+import { OptionPickupModel } from "../models/option/OptionPickupModel";
+import { OptionPricingModel } from "../models/option/OptionPricingModel";
 import { PricingDataProvider } from "../dataProviders/PricingDataProvider";
-import { UnitModel } from "../models/Unit/UnitModel";
+import { UnitModel } from "../models/unit/UnitModel";
 import { LocaleDataProvider } from "../dataProviders/LocaleDataProvider";
-import { ProductModel } from "../models/Product/ProductModel";
+import { ProductModel } from "../models/product/ProductModel";
 
 interface OptionModelBuilderData {
   optionData: OptionData;
