+++ src/builders/ProductModelBuilder.ts
@@ -7,13 +7,13 @@ import {
   RedemptionMethod,
 } from "@octocloud/types";
 import { ProductData } from "../data/ProductData";
-import { OptionModel } from "../models/Option/OptionModel";
+import { OptionModel } from "../models/option/OptionModel";
 import { OptionModelBuilder } from "./OptionModelBuilder";
-import { ProductModel } from "../models/Product/ProductModel";
+import { ProductModel } from "../models/product/ProductModel";
 import { TimeZoneDataProvider } from "../dataProviders/TimeZoneDataProvider";
 import { LocaleDataProvider } from "../dataProviders/LocaleDataProvider";
-import { ProductContentModel } from "../models/Product/ProductContentModel";
-import { ProductPricingModel } from "../models/Product/ProductPricingModel";
+import { ProductContentModel } from "../models/product/ProductContentModel";
+import { ProductPricingModel } from "../models/product/ProductPricingModel";
 
 interface ProductModelBuilderData {
   productData: ProductData;
