+++ src/builders/__tests__/AvailabilityCalendarModelBuilder.test.ts
@@ -1,5 +1,5 @@
 import { CapabilityId, PricingPer } from "@octocloud/types";
-import { AvailabilityCalendarModel } from "../../models/Availability/AvailabilityCalendarModel";
+import { AvailabilityCalendarModel } from "../../models/availability/AvailabilityCalendarModel";
 import { AvailabilityCalendarModelBuilder } from "../AvailabilityCalendarModelBuilder";
 
 describe("AvailabilityCalendarModelBuilder", () => {
