+++ src/builders/__tests__/UnitModelBuilder.test.ts
@@ -1,8 +1,8 @@
-import { UnitModel } from "../../models/Unit/UnitModel";
+import { UnitModel } from "../../models/unit/UnitModel";
 import { UnitModelBuilder } from "../UnitModelBuilder";
 import { CapabilityId, PricingPer } from "@octocloud/types";
 import { UnitDataProvider } from "../../dataProviders/UnitDataProvider";
-import { ProductModel } from "../../models/Product/ProductModel";
+import { ProductModel } from "../../models/product/ProductModel";
 
 describe("UnitModelBuilder", () => {
   const unitModelBuilder = new UnitModelBuilder();
