+++ src/models/availability/AvailabilityCalendarModel.ts
