+++ src/models/availability/AvailabilityContentModel.ts
