+++ src/models/availability/AvailabilityModel.ts
