+++ src/models/availability/AvailabilityPickupModel.ts
