+++ src/models/availability/AvailabilityPricingModel.ts
