+++ src/parsers/__tests__/AvailabilityParser.test.ts
@@ -1,10 +1,10 @@
 import { AvailabilityStatus } from "@octocloud/types";
 import { PricingDataProvider } from "../../dataProviders/PricingDataProvider";
 import { AvailabilityParser } from "../AvailabilityParser";
-import { AvailabilityModel } from "../../models/Availability/AvailabilityModel";
-import { AvailabilityContentModel } from "../../models/Availability/AvailabilityContentModel";
-import { AvailabilityPricingModel } from "../../models/Availability/AvailabilityPricingModel";
-import { AvailabilityPickupModel } from "../../models/Availability/AvailabilityPickupModel";
+import { AvailabilityModel } from "../../models/availability/AvailabilityModel";
+import { AvailabilityContentModel } from "../../models/availability/AvailabilityContentModel";
+import { AvailabilityPricingModel } from "../../models/availability/AvailabilityPricingModel";
+import { AvailabilityPickupModel } from "../../models/availability/AvailabilityPickupModel";
 
 describe("AvailabilityParser", () => {
   const availabilityParser = new AvailabilityParser();
